Move MultiZkvmClient to its own module.
